ENH: Moving Ridge out of preview (#2154)
