+++ onedal/linear_model/linear_model.py
@@ -293,15 +293,15 @@ class Ridge(BaseLinearRegression):
         """
         module = self._get_backend("linear_model", "regression")
 
-        X = _check_array(
-            X,
-            dtype=[np.float64, np.float32],
-            force_all_finite=False,
-            ensure_2d=False,
-            copy=self.copy_X,
-        )
+        if not isinstance(X, np.ndarray):
+            X = np.asarray(X)
 
-        y = np.asarray(y).astype(dtype=get_dtype(X))
+        dtype = get_dtype(X)
+        if dtype not in [np.float32, np.float64]:
+            dtype = np.float64
+            X = X.astype(dtype, copy=self.copy_X)
+
+        y = np.asarray(y).astype(dtype=dtype)
 
         X, y = _check_X_y(X, y, force_all_finite=False, accept_2d_y=True)
